Split up test_settings.py into multiple modules
